new section added

+++ Project/Models/ProductClass.cs
@@ -4,7 +4,7 @@
     {
         public string pname {  get; set; }
         public string ptype { get; set; }
-        public int pprice {  get; set; }
+        public decimal pprice {  get; set; }
         public string username {  get; set; }
         public string date {  get; set; }
         public int isreject {  get; set; }
